Feat: adapt income limits due to law changes
The info message might need another iterationof adaption. Because the
rules have become simpler, the text needs to change more than the raw
number.

The team decided to keep the logic in place that alternatives the income
limit based on certain factors (like single parent). We know that this
separation will come back soon. Furthermore it was the more minimal
change.