Feat: establish limit of simultaneous BEG months
This adaption is based on the related law changes.

So far, this only block the selection of any BEG month when the
conditions are not met. Further UI/UX improvements are planned.

The code is okayisch, but not really nice. The current concept of how
rules are implemented might hit its boundaries.

Notice that another test scenario had to adapt due to the new
regulations.