Feat: hide BEG months in planner that are not selectable
With the introduction of limited simultaneous BEG months, the UX
challenge occurred to support the user with these rules. A first
improvement is to hide BEG months that are no longer selectable due to
this rule.

Notice:
I'm not too satisfied with the current technical implementation. But
there is no good enough reason, nor the time yet to do bigger
refactoring. This change should cause no new technical debt.