Feat: extend settings for elternteile
For the new regulations of restricted simultaneous BEG months, there are
certain exceptions. To properly take these into account, further
information from the form answering process are needed within the
month planner. While the planner has theoretically access to all stores
with all user replies, this logic will spread most importantly into the
domain logic that processes these rules.
Currently, it seems like the domain logic depends on the
`CreateElternteileSettings` to receive all these information it needs to
validate certain rules. Therefore, these settings got extend. While
doing so, the whole data approach could be questioned. In that context,
the current solution must be seen as a minimal extension of the current
concepts. It is neither pretty nor effective.

Notice that this commit does not yet make use of the additional
information in the planning. It purely focuses on enriching the data
set.