Feat: show notification after first simultaneous BEG month
To further improve the user experience, the rule for simultaneous BEG
months should be explained to the user, so she/he does not get confused
why BEG months "disappear". Therefore a notification gets displayed.
This is triggered after the first simultaneous month was selected.

Exceptions to this rule are taken into account and no notification
appears if there is no restriction.

Due to the currently weak domain model, certain logic was exposed (i.e.
exported) to allow to detect this properly. Else, it would have been
necessary to re-implement domain logic outside of the domain, which is
problematic for various reasons. Future refactoring hopefully improves
that.

There have been some minor test case improvement of test cases that are
loosely related.