Feat: use better date auto fixing
The default auto fixing in the date component is not really nice to use.
For example if you type the number `6`, it will insert a `3` instead.
That is because the valid range of inputs are limited to the possible
month and day indexes padded with a leading zero.
If using the `pad` option for auto fixing, it much more intuitive. If
you type `6` for June, it will become `06`. There should be no other
possible month the user could have intended to type. If you start with
a `1` everything is just fine and typing can be continued.