Chore: update react-router-dom dependency
Though no major version, there was some kind of breaking change
included. At least from a testing perspective.
The `useNavigation` hook got a strong dependency to the router context.
To make the tests of affected components work, this context needs to be
provided. In turn, mocking `react-router` does not work that easily
anymore.

The currently existing tests are mostly unit tests, or smaller
integration tests. None of them include actual routing yet, but rather
mock it away. As mocking has become hard to do, the question of the
exact testing scope has been raised. In that context we decided that the
routing (and its effects) should not be tested on page component
level. But rather it will be tested on a bigger scenario tests in an
end-to-end setup. In such a setup it is possible to use proper user
centered testing without relying on a mocked routing that checks
specific API calls.

As we don't have the end-to-end test setup yet, we removed some of the
tests completely. For some it was fine to remove certain expect
statements. Anyhow, a ticket was created to remind us of these test
scenarios which are no more covered. Though it could be questioned how
well they were covered so far.