Chore: improve a11y checks and fix a violation
The `create-react-app` setup enables per default only a small set of
a11y rules to check. But there are many viable rules which should be
checked if a11y is a serious topic. The documentation points out how to
opt-into checking for all rules.

Positively, only a single violation was detected with the additional
rules. A hidden checkbox was abused with a label as stand-in component.
Using an actual interactive element like a button solves the issue
accurately.
Respective event handlers are necessary as an "aftermath".