Test: improve critical change month tests
The tests for the `changeMonth` function are quite crucial as this
logic represents an important part of the domain. Therefore it is of
high priority to have a good test suite for it. That includes some clean
code with easy to write and, most importantly, easy to read test cases.
This helps the developer to work comfortably with the tests and support
test-driven-development.
The added wrapper/test-harness reduces the average number of lines per
test case by >50%. Plus these lines are easy to grasp. At the same time,
the original source code remains untouched for now to avoid big
refactorings. Therefore this solution is tailed for test purposes.
More details on this on the test suite itself.