Chore: use strict engines settings
Reproducible builds and a strong toolchain is very helpful for daily
development. Using a fixed version for NodeJS and NPM is a good practice
with native support by the tools themselves. So far this was already
defined, but very loosely. For example does our `engines` requirement of
NodeJS above version `16.14.2` not even satisfy the `engines`
requirements of our `@bmfin/steuerrechner` dependency (threw a warning).

To fix the latter and follow a more strict policy, the version has been
bumped to the latest long term support version (which is also sufficient
for the mentioned dependency). Furthermore NPM has been configured to
cancel any command if the engine requirements are not matched.

The already in-use GitHub action to setup NodeJS does support this
natively and reads the `engines` settings. So no duplicated
configuration necessary here.

The topic of a unified developer/CI toolchain should be discussed in
future.