Chore: fix dependency issues
There was a vulnerability issue with the `webpack-dev-middleware`
sub-dependency. This originates from the unmaintained `react-scripts`.
To fix that, the parent/intermediate dependency `webpack-dev-server` has
been bumped manually. Unfortunately this lead to breaking changes
causing issue to run the development server. In fact the development
server simply did not start anymore.
To fix that, the `webpack-dev-middlware` was now manually bumped to
a patch version that fixes the vulnerability. Else this is the same
minor version as by `react-scripts` defined.