Refactor: add default option into CustomSelect
So far, every component that was using `CustomSelect` input component,
added an additional (first) entry called "Bitte wählen". That is quite
cumbersome and annoying to do. Therefore this has been integrated into
the select component itself.

Furthermore it has been improved to
actually work as a placeholder option. That means for example that this
option is not visible in the open selection, nor selectable when this
selection is required. Notice that is is still available for not
required selections as the user needs an option to undo a possible wrong
selection.

This is a preparation step to also allow selections with options values
that are no strings.