Fix: show simultaneous BEG warning only once
This feature was lost due to missing tests during the law change
adaptions. The idea was to show this warning only when the first
simultaneous month was selected. Unfortunately it popped up all the time
once this simultaneous month was selected. Though it show again in case
these months get unselected and again.

This has been fixed very frankly and in a quick live coding manner with
the stakeholders together. Time is pressuring due to the holidays and
the go-live. Since we'll focus on the new planner anyhow, this shouldn't
be too much dept for too long.