Fix: fixed months for financial inputs
This is a logically more involved issue. The assessment period is not
static and fixed to the consecutive 12 months before birth. But a group
of 12 months before the date of birth with possible gaps in-between. It
is possible that some months are not taken into account, so months
further in the past are considered. At typical example are maternity
months.

In result, the from inputs to define financial income related to the
assessment period need to be more flexible to reflect this. As we
currently can't redesign this whole input flow, a quick fix has been
established. Thereby we just count the 12 months. It is the cognitive
task of the user to map these to accountable months. This is not a great
user experience, but from the domain logic perspective, the bug is
fixed.

The whole time period logic in the legacy code is quite cumbersome.
Especially since there is already the month counting logic for financial
input after the date of birth. The current logic encodes and decodes
everything to and from strings, guessing between them and compare pure
strings for sorting etc.
An attempt to resolve this completely, get full type support and remove
the support of raw dates has failed. At least for a fair amount of
working hours with the current focus point. Thereby, it has been decided
to just use the month counting support that is already there and apply
if also for the months before birth.