Style: remove custom reboot styling
The styling of the application is currently a mess and it is really hard
to analyze what applies where. Especially since everything is globally
just with BEM here and there.
One issue is the repetitive normalizing/rebooting. There is the reboot
of the surrounding Familienportal on production. There was the (non
functional?) `normalize.css`. There still is the reboot by Bootstrap and
before this commit there was some additional custom reboot too.

The removal of that stylesheet actually didn't cause too much trouble.
Though visual checks have been performed manually. There are a few
elements like paragraphs and ordered list or anchors which gets some
weird (and definitely not normalizing) paddings. Most of them are not
noticeable or just okay. Where not, this has been fixed locally now.
Thereby we avoid the problem on a global level and shift it towards the
smaller components, which is easier to handle on the long run. In fact,
the Bootstrap integration is there for a reason. So if there should be
exceptions, they must be exceptions. Anyhow, this is a future topic
anyhow.