Style: inline hide-on-desktop utility
This commit lives in the context of getting the styling of this project
under control.
While utility classes are a great approach, media queries based on
screen sizes are not. They should be absolute exceptions, if at all.
Thereby, in the above mentioned context, this utility has been inlined
locally to the components using it (only 2). This helps to clean up the
global scope and slowly get into a refactoring position.

The `mixins` style directory has been deleted as it was just empty
anyhow.