Style: remove Bootstrap reboot styling
As described in the commit before, the normalizing/reboot of styling is
redundant and confusing. This time the Bootstrap reboot has been
removed. This allowed to revert some of the fixes that where necessary
in the last commit. Bottom line: basically zero code necessary.

Also some weird spacings which occurred before (but where classified as
not harmful) got solved by this.

The whole global styling space has become much more narrow and easier to
argue about.