Style: get rid of remaining Bootstrap parts
This commit lives in the context of getting the styling of this project
under control.
Unfortunately we have no access to the former VCS history to retrace
what actually happened. But it seems like Bootstrap has been used much
or phased out. But not completely. And certain side effects and styles
where still taking effect. Though not many.
Removing bootstrap finally all together leaves the styling of the
application again more simpler and more predictable.