Style: take fonts from angie package
This adds the `angie` styling package of the company. This is the first
of two such packages. So far this is only used for the fonts. It allows
to get rid of various files and definitions with a more concise result.

Notice that the `font-size` variable is still an issue for now. There
are some calculations which are based on it and it is not easy to get
rid of it. The whole clean base font-size is still an open topic.