Style: initial setup of TailwindCSS
This already includes the agreed on custom spacing, colors etc. which
are agreed on with design. They are based on the old values of the
application. Everything is kept quite close to the documentation in
Figma.
This still uses the old colors for the Elterngeld varianten. But it uses
already the new color names for them.

Notice that the during SCSS compilation, it complains that you can not
use `@tailwind` before `@use`. But putting it after will load the
normalizing in a later stage, removing all the typo styling loaded
earlier.
There seems to be no "proper" solution. So I just followed the
installation guide to add Tailwind in the `index.css` instead. That is
not great, but also not horrible.