Chore: add material design icons
The application does not incorporate any icon library yet. Instead is
maintains its own implementation with SVG paths. Our style guides
heavily recommend the usage of the Material Design icons. These are also
used in our designs.

As the company has it's own comfy distribution of the icons, that one
is used as dependency.

This does not yet resolve the old icons and only adds the integration.
Thereby enabling the proper development of new components/code. The
former is task of a new refactoring ticket.