Feat: add new Elterngeldvarianten description
The descriptions of the Elterngeldvarianten is target for optimization.
This includes a new design and functionality. Therefore this commit
takes the first minor step towards that direction.

The custom set `DEFAULT` border radius in the TailwindCSS theme is
target of conversation with design. It aligns with the current designs,
but might become part of a more detailed scale.

Note:
Certain best practices in code are still unclear to the team and might
change (e.g. component-file relationship).