Fix: misaligned button content
Due to the new CSS normalization which comes with TailwindCSS, the
Button component was causing awkward boxing of its content. It is
a little weird, but unfortunately it is unclear what causes the issue in
exact.
As the normalization is useful in general, and the issue is that
focused, it has been solved by some styling attribute on the Button
component itself.