Feat: add payout amounts for variant descriptions
The calculation of the payout amounts is quite involved in the old code.
Because this calculation needs a lot of parameters based on the user
input on earlier pages, the calculation is currently quite tangled into
the Redux store(s).
The direct usage from the store comes with side effect to the components
the store was designed for. Like the strict coupling to additional
income values and the caching of asynchronous results.
The taken approach tries to minimize the footprint in the context of
having no opportunity to refactor the whole calculation logic and how it
is integrated. All the "weird" parts are encapsulated into the
`usePayoutAmounts` hooks, single purposed for the usage of the variant
descriptions.

For probably historic reasons, the calculation is asynchronous. The
current speculation is that this is because the former usage of a remote
endpoint for tax constant retrieval. Which is still supported by the
code, but by current judgment no more used. Anyhow, this should just
give an insight why the asynchronous effect on initial component render
is used. There should be no reactivity in this case. If users make
changes to their inputs that have an effect on the calculation, they are
required to navigate away and back to the descriptions. Which will
trigger the calculation just again.

I'm seriously not happy with the testing. Just as a statement here.