Style: add scale for font sizes
So far we relied on the pure typo styling of HTML components. But it
turned out with pure cascading and inheriting, this does not work for
the designs. So we must be able to manually set the font-size for
certain elements.
The scale naming is again based on numbers related to the pixels on
standard font-size (still 18px atm.). Only those which were defined in
the style guide are set here.