Style: move TailwindCSS loading to separate file
When debugging styling, it is important to know where certain styles
affecting an element are coming from. Having a broad `index.css` is
thereby not really helpful. This is especially true for the integration
of styling libraries. This especially applies when they include some
kind of resetting.
Having a dedicated `tailwind.css` stylesheet helps with that (as this
exact name appears in the browsers development tools).

Notice that this also introduces the `styles` directory. This is a quite
common practice. Existing directory names mostly follow these
best-practices. The still existing `src/globals/scss` is now limited to
former SCSS files and is about to get phases out slowly.