Fix: remove unintended color references
Some styling was referring to the color variables of the parental
allowance variants where this was semantically not correct. Like the
background color of the menu. But because some colors have the same
value, this has no impact for the user. For example it should actually
be the primary color, which happens to be still the same as the one for
Basiselterngeld.