Fix: padding of list element
In HTML + CSS, the "dots" of list items do not belong to the content box of the
list. That means if there is a left padding, they show up within that
padding area. This can be resolved with inline list styling, but that
has other undesired side effects here.
This is a difference between the Figma design tool and web technologies.
Thereby the horizontal padding must be adapted to achieve the same
visual outcome.