Style: resolve padding function and shim down
The padding function for SCSS was no good practice where an outline
should be used. In any way, we can revisit this styling if it should
become an issue with the design concept.

As another step, the whole `functions` folder of the SCSS directory were
just two simple functions, getting re-exported. To further reduce the
footprint of the old global styling, this has been shimmed down into
a single file.