Fix: usage of fonts in preview build
The Bundes* fonts where no longer loaded on staging with the preview
build. This remained undetected for a while because it works in
development mode and there are no specific (visual) tests for this.

Since we changed the font application a little with the `angie` package,
the `woff` files have to be copied over for the preview too. This
requires to follow the full paths of the build.

Notice that the font topic seems to be a little messed up due to the
`public` folder which also integrates the same fonts, but uses different
paths/names. We should ensure that the full `static` folder is actually
working for the next deployment to the Familienportal.