Style: change colors for variants and adapt
This introduced the new colorscheme for the parental allowance variants.
The former `-light` colors could be removed.

In result a few components displaying content related to the variants
had to be adapted in style to have them still work. This was done in
pairing with design together. Though it might receive another iteration.