Fix: build of preview
Recently the preview build created some issues when the new font loading
was introduced. And while this was fixed at first, it does not work in
our staging environment with a base path on the URL the application is
served from.

The solution uses a much simpler approach than the one before. It
remains the `static` output of the application build untouched and does
a direct linking into the `index.html`. In fact it now works the same as
for our development `public/index.html`.
No, any base path just works fine as we have exclusively relative paths
and do not touch them.
It is also arguably closer to our "production" build where we send over
the `static` folder content of the build.