Chore: detect and remove unused or dead code
The project suffered from a good amount of unused export statements and
in result dead code. This is an issue for any TypeScript project and it
is a part of code quality assurance to detect this continuously.

The application code base suffers from more dead code, which is not
easily detectable by automated tools. These are for example the time
period with dates, BMF variables via HTTP request, premature babys, etc.
Getting in control of what is automatically detectable gives us some
grip.

This does not yet run `knip` in production mode. This is in first place
because some patterns still expose code due to testability, leading to
false positives. This might be re-visited as a team later. Because this
way we might miss out relevant positive matches (e.g. only used by
tests).