Chore: further use of list vs. ET1;ET2
Many data structures use the keys ET1 and ET2 for the both parents.
While already not the best naming in general, this falls short quickly
when there only a single parent. And in potential future cases when more
than two.
It seems like using lists work quite well for many data structures. At
least for the UI components that have to deal generically with different
number of parents (1 or 2 atm.). In the domain language there is still
the usage of parent one and parent two, as this is the language of the
law etc. Though, certain aspects of the application, specifically the UI
probably can profit by making zero assumption about how many parents
there can be. And as this makes them more powerful, plus more simple,
there seem to be zero downsides to this approach where applicable.