CI: separate license auditing
With the current pipeline and possibilities to deploy, it is unclear
which CI steps should actually be blocking. Like any pipeline usually
have some which are not. Having a single job makes it hard to establish
such steps.
At the moment, in the current state of the project, it is unclear how
to continue on licenses. Therefore it got separated for now.

In general, the team started the discussion how to continue on improving
and iterating the pipeline. A split into more jobs was part of that
discussion.