Chore: remove Knip temporarily due to license
The license issue is not directly with `knip` itself, but a three level
down dependency called `deep-freeze`. It is unclear how to follow up on
this for now.
There are ongoing conversations with the author of the intermediate
dependency `@ericcornelissen/bash-parser` to remove this dependency at
all, which would resolve that issue.
Else, we try to clarify the usage of the `PUBLIC DOMAIN` license for our
circumstances.