Fix: summation footer wrapping after theme change
An issue got undetected when we shimmed down the spacing scale in the
styling theme. As there is no proper linting (yet) to detect wrong
utility classes, this was forgotten.
But as we have no longer this big spacing, I was in need of another
unit. I feel a bit uncertain how to deal with such cases yet. I'm not
too happy with the arbitrary unit, but it works well for now.