Style: extend SVG axioms for integration into text
As already mentioned for the fill color, SVG icons work best when
integrated into their surrounding text by default. That also includes
the sizing, alignment and spacing.
The current sizing and alignment is implicitly based on current design.
But it will need further refinement with design together.

In result a couple of usages could be simplified. The alignment comes
naturally, the sizing as well. And simple whitespaces do proper spacing
instead of tweaky gaps/margins.