Chore: permit pervasive relative imports
Massive use of relative imports are no good practice, especially towards
the parent level. Imports are really hard to read, write and they are
more prone to moving source files. Relative parents within the same
directory are just fine and again more stable.

This sets up a path alias for TypeScript that helps to write absolute
import paths. In combination with the new linting rule this becomes
automatically fixable.

Notice that the `import/no-relative-parent-imports` rule of the already
existing `eslint-plugin-import` package does not fully do the same, nor
is it automatically fixable.