Test: fix path alias for tests import resolution
Unfortunately this gets complicated as CRA does not allow to configure
this easily. I have found no other way than to opt for CRACO which is
made to solve these kind of problems without fully rejecting CRA.

Notice that it is still on the agenda to migrate away from CRA due to
various reasons. This adds up on these.