Feat: enable partner months conditionally
First of all: I hate my work in this commit with every cell of my body.
We seriously need to refactor the central logic of the planner. I tried
to do something here and there, but it is pervasive and not really
possible to follow the scout concept here.

There is the concept of additional partner months. Single parents always
gets these additional months without further conditions. Else, both
parents need to take at least two months of basis or plus months. Only
then the additional months are available.
In the past this was only handled by hints and notifications to inform
the user. But in the end it was possible to create a wrong plan. This
change now ensures that the additional months are only enabled when all
conditions are met.