Feat: add user tracking setup
We finally wanna track some additional user behavior to drive out
development with data. Because the Familienportal we are running inside
uses Matomo, we integrate into their tracking. But we setup a custom tag
manager container that we load ourselves.
It is very important to load this container only when the user has
confirmed the cookie banner of the Familienportal! The actual tag
manager to load is configured via a build-time variable.
The general setup logic follows that of the official Matomo
documentation. Just a bit more clean and readable, instead of squashed
JavaScript code.