Feat: enforce correct usage of partner months
First of all: I hate my work in this commit with every cell of my body.
We seriously need to refactor the central logic of the planner. I tried
to do something here and there, but it is pervasive and not really
possible to follow the scout concept here.

There is the concept of additional partner months. Single parents always
gets these additional months without further conditions. Else, both
parents need to take at least two months of basis or plus months. Only
then the plan is valid.
In the past this was only handled by hint and notifications to inform
the user. But in the end it was possible to create a wrong plan. This
change now ensure that only valid plan can be created. Therefore it
disables the selection of further months for a parent when its limit is
reached, even if there is a contingent left. These remaining months can
now be only selected by the other parent.

I intentionally did not add further tests to the `Monatsplaner`
component. Testing this feature through this highly integrated component
is incredibly cumbersome at the moment and does not seam to help with
reliability. As we very soon completely rewrite this whole art of the
application, it currently would be wasted working time. Though, this is
of course not a good practice at all.