Feat: include EG+ into partner month limits
Again, for the protocol: I hate myself for these commits.

The just added feature to restrict month selections in a plan according
to limits by partner months, the Elterngeldplus was not taken into
account. The actual limit per partner must be determined by the combined
selection of Basiselterngeld AND Elterngeldplus together. This makes the
whole ruleset quite "nasty" to work with. There are no edge cases due to
the ratio of Basiselterngeld to Elterngeldplus.

The testing of these rules have been mostly pushed into the logic of the
`change-month` domain. That's because the testability via the
`Monatsplaner` component becomes incredibly cumbersome and hard to
understand/maintain. But more importantly we will soon completely
redesign this part of the application. So testing further down gives us
more confidence.