Chore: add markers for tag manger tracking
The current approach for user tracking configuration is to use a tag
manager. This should enable our project managers to participate better
and potentially take over. Thereby we do no tracking via code directly.

But unfortunately does the tag manager of Matomo not support any
user-centric triggers. So for example visible elements are identifiable
by CSS class name(s) or identifier only. Therefore it is sometimes
necessary to add such markers to the code to provide possibilities to set
triggers. That is not optimal and not beautiful.
But it is a compromise for now. It is important that we start to do
tracking early and shift to more robust/smart/... triggers in iterative
steps.