Fix: remove needless random keys
This was "discovered" as it was reported by the code quality scan as
a security hotspot. It was actually about the save usage of
`Math.random` from a cryptography perspective.
But actually these random keys are not necessary at all. Just to be
sure, I checked manually if the transmission of data to
ElterngeldDigital still works and everything is fine.