Chore: enforce readonly properties
This has been raised as code smell by our code quality scan.
Theoretically this doesn't affect as, because we also enforce the
spreading of properties, which copies values in memory anyway. But we
don't want to establish unnecessary rule exceptions if not necessary.
This code smell can be detected and automatically fixed by our linting
rules. Therefore the only downside might be the readability due to many
`readonly`s in the code. Furthermore it might become relevant for us in
future.