Feat: integrate partnerschaftlichkeit into planner
It was a "tricky" question where to put the integration. A direct
integration into the store might have been a valid alternative as it is
owns the relevant data and makes it component independent. Though there
is no obvious mechanism available to do this action independent.
Currently the only component changing the data is the planner. And we do
not intend to change that. But is it store approach is under review.
Thereby the current approach is okay, but might be iterated.