Feat: add tracking for 'nutzergruppe' metric
Originally this was implemented within the tag manager of our tracking
service itself using a custom JavaScript function. That does no more
work well for us, as it requires administrative permissions we don't
have.

This basically just takes the value of the user inserted date of birth
and maps it to one of our three defined user groups. Therefore it
compares it with the current date and puts the result onto the tracking
data layer. From there it can be read by any tag of the container.
The data is accessed and written in a way that makes no assumption when
and by which trigger the data is read. Making it available as soon as
the user provides some input.