CI: add custom action to deploy on GitHub pages
The idea is to have multiple versions of our page being served
simultaneously. Therefore, we use the option to serve from a branch
instead based on build artifacts. This way we can create (nested)
directories with different versions. Knowing the path allows it to be
attached to the base URL to view this version. Per default the version
of the `main` branch gets shown as before.

Further logic to remove versions again is still needed and will be added
in a later step.