CI: fix fetching during deployment
A plain `git fetch` does not work because the repository is private.
Therefore the initial checkout action is used to do a deep fetch. That
is not optimal and need to improve in future for short pipeline runs.

I'm afraid the `git push` will also not work for the same issue. Let's
see.