CI: remove base element for deployments
Apparently there is a difference between the local test setup and the
actual GitHub pages service. Locally, it was necessary to add a `<base>`
element for the path of the version. This was especially the case
because the fetched sources from the Familienportal use absolute
references. On staging this leads to double paths. So if there is `<base
href="main/" />` so all sources under `https://<base-url>/main` work, this
leads to network request of `https://<base-url>/main/main/<resource>`.