CI: do not deploy if there are on changes
Currently the pipeline actually fails if there are no changes to the end
result of the to deploy page. That's because this would end up in
a empty commit. This can happen for example for changes to the CI
pipeline. ;D

Also includes minor improvements in job naming for readability on
GitHub.