CI: inline deployment into related workflow
The idea is to keep things more tightly coupled (e.g. permissions) and
rather slim. Therefore, the idea is to use a workflow only. It is
possible to call workflows from other workflows and maintain multiple
triggers, like the manual one.

This also tries to "clean up" to script a little by using more
environment variables, separating logic and configuration.