CI: integrate re-usable workflows
This is the next step which removes the former action and makes
a workflow call instead. It is intentionally not a `workflow_run`
relationship to make it more explicit and also to have a proper input
available.

To make this work properly, the deployment report happens within the
deployment workflow. Which seems obvious.