CI: conditionally build only on deployment
To better/faster integrate into other workflows it is possible to define
an input as an artifact name that will be downloaded for deployment.
Therefore, the workflow builds and uploads an artifact itself when no
such input was provided. The consecutive job then downloads either the
input asset or the own one.
In the end it always ensures that there is an artifact of what has been
deployed.