CI: re-use build for main pipeline
This actually makes use of the added feature of the deployment workflow
to skip the build and use an existing artifact. This shortens the
pipeline duration again, as this does not re-run the build step again.

Side note:
I opened a ticket actions/download-artifact#334 to better link upload
and download artifact jobs/steps.