CI: separate concerns of workflow and shorten
The spaghetti steps of the `basics` job is a little inconvenient when
tracking runs and observing run times.
Also having the build run in parallel to the linting and tests, it
should shorten the runtime by more than a minute.

Also creates a tiny bit of better relationship between build and deploy
jobs using outputs and inputs. Though, there this is still the open
ticket to make it tightly coupled.