Feat: add hook to gather data for summary page
Data transformation layer to translate between different data models. We
already have a plan how to reduce this load in future.

This starts to make use of `immer` dependency for easily create states
that are easy to read in tests. It was already a child dependency due to
the Redux toolkit. Though, has become a first level dependency now.