Feat: shim down former summary page to button
The old page gets replaced by a new design. But there is still the
feature that allows to transfer data to the actual application form.
This was implemented by a hidden form with submit button. This lonely
part of the old component is still relevant. So the hole component got
reduced down to this only purpose. All child components, styling etc.
pp. got removed by that.

Long term, this feature is imagined to work not as a component but
rather a pure function or React hook. But this is one refactoring step
further and costs some issues. Though it will be worth it. Anyway, step
by step.