Feat: remove asterisk for required input fields
User testing has pointed out that this confuses users. Almost all of our
input fields are required if visible. The only ones that are not,
explicitly state that they are optional. If a required input is missed,
the validation fails and points the user to the input. Therefore we
decided to just remove the asterisk completely.