Feat: improve text for Planungsübersicht
Project partners noted that `Bonuskelterngeld` is not a proper notation
(we updated our glossary).

Consistent usage of `Einkommen` instead of `Gehalt`. Due to that, remove
second line of table headers. It was unclear how this now works from the
wording. But we decided that this line might not be helpful at all.