Chore: exchange CRA for Vite as bundler
This was a long standing topic as create-react-app is unmaintained for
a while already. Which, among other problems, caused vulnerability
problems. We also had to add CRACO to solve some shortcomings.
More information about the decision in the related
architecture-decision-record.

Notice that this is an incomplete commit. Should mean that the
repository is partially broken by now. While building etc works, linting
and testing for example does not. Thats because create-react-app was not
just a simple bundler.
To keep the commits of reasonable size and easy to understand in future,
the transformation work got split into multiple commits.

This sneaks in the removal of the remaining JavaScript source files.