Chore: exchange Jest with Vitest as test runner
The former CRA "bunder" also fully configured the test
runner/environment. Instead of sticking with Jest, we decided to switch
to Vitest instead. Checkout the architecture-decision-record for more
details.

Most changes of this commit are just a simple switch of Jest to Vitest
mocks. As the Vitest API is almost the same as of Jest, this was in most
cases very straight forward.

This removes the settings of longer test timeouts. They seem to be not
necessary, at least on a global level. There is a single test that runs
long and it has been configured individually. Thereby, new long running
tests should be detected immediately and can be fixed early. In best
case the current one will be so too.