Chore: unfold ESLint configuration
With the removal of create-react-app, the included ESLint plugin and
configuration sets are missing too. There it was necessary to manually
get back all the hidden configuration, taken from the obsolete plugin.

There were not 1:1 replacements for all parts. In result the linting
with the more up-to-date dependencies got a little better and discovered
issues, which remained undetected so far. Though, none of them was
critical in any way. Anyhow, more clean code for sure.

The Vitest globals configuration should not be necessary. It is just an
easy fix for now. In general the ESLint configuration should be
re-written soon, using the new configuration format and better
TypeScript integration.

Furthermore there was some need to fix some ESLint file matching/ignore
patterns. This is a little confusing and annoying. But with the upgrade
of ESLint and the configuration system ahead, these problems should
resolve itself.